Add non-sql dependency attribute
Resolve-dependencies between fields are not in the
graphql-specification, still several libraries as graphql-resolvers or
apollographql/graphql-tools solve this by introducing new methods. The
SQL-approach join-monster does not offer this functionality, although it
would be easy to offer access to siblings, after these were resolved by
join-monster.

One can request to load other table columns using the 'sqlDeps'-field
and to process the fetched data in a sibling field. The limitations for
this are, that the sibling field is in the same table. More
sophisticated post-processing would have to take place in the query's
resolve method, traversing the resolved object and manipulate it, which
is highly discouraged, as it undermines graphql's type system.

This commit adds a new possibility to list a sibling field, that is a
GraphQLObjectType, or a list of those. When building the sqlAST, the
object in the 'directDependency'-attribute will be added to the SQL tree
and resolve as if it was in the query. Later, the depending field can
access it implementing a resolve method.